refactor for bloggers api

+++ src/repositories/blogs/blogs.repo.ts
@@ -1,9 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable } from '@nestjs/common';
 import { Blog, BlogDocument } from './blogs.schema';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { BlogCreateInputModelType } from '../types/input.models';
-import { CreateBlogDto } from '../types/dto';
+import { BlogCreateInputModelType } from '../../types/input.models';
+import { CreateBlogDto } from '../../types/dto';
 
 @Injectable()
 export class BlogsRepository {
@@ -13,9 +13,12 @@ export class BlogsRepository {
     await new this.blogModel(newBlog).save();
   }
 
-  async deleteBlog(id: string): Promise<boolean> {
+  async deleteBlog(id: string, userId: string): Promise<boolean> {
     const blogInstance = await this.blogModel.findOne({ id: id });
     if (!blogInstance) return false;
+    if (userId !== blogInstance.userId)
+      throw new HttpException('Not your own', 403);
+
     await blogInstance.deleteOne();
     return true;
   }
@@ -23,9 +26,12 @@ export class BlogsRepository {
   async updateBlog(
     id: string,
     inputModel: BlogCreateInputModelType,
+    userId: string,
   ): Promise<boolean> {
     const blogInstance: BlogDocument = await this.blogModel.findOne({ id: id });
     if (!blogInstance) return false;
+    if (userId !== blogInstance.userId)
+      throw new HttpException('Not your own', 403);
     blogInstance.updateBlog(inputModel);
 
     await blogInstance.save();
